# packages/conversion-service/swagger.json
{
	"components": {
...
		"schemas": {
			"IConversionProcessingResponse": {
				"properties": {
					"conversionId": {
						"type": "string"
					}
				},
				"required": [
					"conversionId"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"conversionId": "55309c37-aed3-4ee9-a143-f1e305333189"
				}
			},
			"IUnsupportedConversionFormatError": {
				"properties": {
					"errorCode": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"EConversionStatus": {
				"enum": [
					"converted",
					"erroneous",
					"in queue",
					"processing"
				],
				"type": "string"
			},
			"IConversionStatus": {
				"properties": {
					"conversionId": {
						"type": "string"
					},
					"sourceFormat": {
						"type": "string"
					},
					"targetFormat": {
						"type": "string"
					},
					"path": {
						"type": "string"
					},
					"retries": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"$ref": "#/components/schemas/EConversionStatus"
					}
				},
				"required": [
					"conversionId",
					"sourceFormat",
					"targetFormat",
					"path",
					"retries",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "conversion-service",
		"version": "0.2.0-0",
		"description": "A webservice with a REST Api which can convert document/media files using different binaries",
		"license": {
			"name": "GPL-3.0-or-later"
		},
		"contact": {
			"email": "sven.patrick.meier@team-parallax.com",
			"name": "Sven Patrick Meier",
			"url": "https://team-parallax.com/svenpatrickmeier"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/conversion/v2": {
			"post": {
				"operationId": "ConvertFile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IConversionProcessingResponse"
										},
										{
											"$ref": "#/components/schemas/IUnsupportedConversionFormatError"
										}
									]
								}
							}
						}
					}
				},
				"description": "Adds the file from the request body to the internal conversion queue.\nThe files in queue will be processed after the FIFO principle.",
				"tags": [
					"Conversion"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"conversionFile": {
										"type": "string",
										"format": "binary"
									},
									"originalFormat": {
										"type": "string"
									},
									"targetFormat": {
										"type": "string"
									}
								}
							}
						}
					}
				}
			}
		},
...
		"/conversion/{conversionId}": {
			"get": {
				"operationId": "GetConvertedFile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IConversionStatus"
								}
							}
						}
					}
				},
				"description": "Returns the current status for a conversion given a conversionId",
				"tags": [
					"Conversion"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "v2",
						"required": false,
						"schema": {
							"default": false,
							"type": "boolean"
						}
					},
					{
						"description": "Unique identifier for the conversion of a file.",
						"in": "path",
						"name": "conversionId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/conversion/{conversionId}/download": {
			"get": {
				"operationId": "GetConvertedFileDownload",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							},
							"application/pdf": {
								"schema": {
									"type": "string",
									"format": "byte"
								}
							}
						}
					}
				},
				"description": "Returns the current status for a conversion given a conversionId\nIf status is 'converted' the resulting converted file\nwill be available to download.",
				"tags": [
					"Conversion"
				],
				"security": [],
				"parameters": [
					{
						"description": "the id of the file-conversion that is requested",
						"in": "path",
						"name": "conversionId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "http://localhost:3000/"
		}
	]
}